+++ src/main/java/camp/nextstep/edu/kitchenpos/tablegroup/dao/TableGroupDao.java
@@ -1,6 +1,6 @@
 package camp.nextstep.edu.kitchenpos.tablegroup.dao;
 
-import camp.nextstep.edu.kitchenpos.tablegroup.model.TableGroup;
+import camp.nextstep.edu.kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
